exit.c: ft_atoi typecaste into unsigned int // added check to complex exec

+++ src/built-ins/exit.c
@@ -6,7 +6,7 @@
 /*   By: ischmutz <ischmutz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/09 11:02:30 by ischmutz          #+#    #+#             */
-/*   Updated: 2024/05/05 18:26:18 by ischmutz         ###   ########.fr       */
+/*   Updated: 2024/05/06 12:39:47 by ischmutz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -103,7 +103,7 @@ void	ft_exit(t_bigshell *data, t_command *cmd)
 	}
 	if (cmd->arg_num == 1)
 	{
-		update_exit_stat(data, ft_atoi(cmd->args->str) % 256);
+		update_exit_stat(data, (unsigned int)ft_atoi(cmd->args->str) % 256);
 	}
 	exitcode = get_exitcode(data);
 	free_struct(data);
